Fixed resize window ops after drag-and-drop of video
If the user runs the .exe directly, an empty window is display onto which they can drag-and-drop a media file. In this case, UI.Initialized isn't set in formMain.FormCreate. If they then dropped a video and resized the window, the aspect ratio wasn't being correctly adjusted in UI.formResize because UI.Initialized was still FALSE.