Possible fix for Create WIC Imaging Factory issue on CreateMessageDialog
This error appears after doing several deletes in the same instance of the application. No idea why it doesn't appear immediately but only after it has displayed several confirmation dialog messages. This seems to have fixed it but I really need to find more videos to delete to test this fix more.