pre-empt a video's window size
The resizing loop is mostly a leftover from the WMP-based incarnation of this app.  UI.setWindowSize now does a better job of pre-empting the maximum size of window required which allows for the screen height whilst also maintaining the correct aspect ratio to fit the video exactly to the window. This helps reduce a lot of jumping around of the window, especially when switching between TikTok-type videos of differing sizes.