Reverse previous change
There's more to this than expected.
Occasionally, mpv doesn't correctly notify the client of the end of a video, so this app doesn't then autoplay the next file. This was fixed by issuing mpsEnd within 0.25 of a second before the end. However, mpv continues playing the final 0.25 seconds and then issues another mpsEnd. This makes this app skip a file. The solution needs more work than this quick fix.