Fix audio cover art. Auto-centre mouse cursor.
For audio and videos files, if auto-centering is on, when a new file is played, the cursor will be automatically positioned in the middle of the progress bar for easy clicking.
Fixed the audio window sizing so that cover art is displayed correctly or, if there is no cover art, the minimalist "letterbox" audio window will be displayed.