Allow for screenshot location in mpv.conf
If no location is specified, MPV defaults to saving screenshots of videos in the same folder as the video. However, for screenshots of images it seems to default to Windows\System32!
I over-compensated for this problem in v2.0.3 by forcing all screenshots to be saved in the current folder, forgetting that the user may have specified a location in mpv.conf for video screenshots.
This fix checks first before setting the location.