auto-download latest release
The about box now displays the latest available version.
If the version of the running app doesn't match the latest available version, the latest release will be downloaded. It won't yet be installed. This will all need refactoring but this is working code. The file won't be downloaded if it already exists. A draft download window shows the progress of bytes downloaded.