Implemented release notes
Downloads release notes for the latest release and any images in them.
Added a What's New? button to the about box.
The release notes form is currently displaying a hardcoded release notes file.
Had to rename consts.pas to mmpConsts.pas as it clashed with vcl.consts in htmlPrinter.pas and caused markdownViewer to fail to compile.