smart [G] and Ctrl-[G]
Holding down [G] or Ctrl-[G] will a) expand the thumbnail window according the maximum number of thumbnails that can be viewed horizontally and vertically, or b) shrink the window to one thumbnail.
Also, fixed a bug where the Stream List kept reverting to the first tab after the user had selected the second one!