Fixes for normal-sized monitors
This code was a left-over from when the app was built on Windows Media Player and had to jump through all sorts of hoops to size the window properly because there was about a 3 second delay before the WMP ActiveX could report the video dimensions. The loop I put in works fine on my fast machine with a large monitor but causes problems I didn't know about on smaller monitors, such as a typical laptop. Therefore this code had to go. I'm currently making lots of changes to fix display issues nobody reported to me.