tweak to segmentAtCursor
can't decide if the original could ever be wrong (the second half is true of every segment to the right of the cursor position, isn't it?), but it's a strange way to code it, nevertheless. It read funny, so I've changed it to more traditional wording. Now it definitely reads correctly.