implement allowIntoWindows setting
Also, fixed spurious file extensions being allowed into the playlist, e.g. .h because it matched with .hdv (and others). The previous issue with '.avi' matching with '.avif' should have been a clue! We now add a space to the file's extension and to all extensions in mmpConsts so we no longer get partial matches.