Completely re-architected as an MVVM app
Or more accurately, V-VM-M (view, viewModel, model).
Employs a notifier/subscriber messaging mechanism to achieve loose coupling of units and objects.
Mostly there but still lots of testing, code rejigging and tidying to be completed prior to a v4.0.0 release.
Timeline editing and the Image & Thumbnail Browser have been left largely as they were: there was no value in breaking the code just for the sake of it.