Fixed edge case problem of videos with decimal duration
When switching to the next video with the Timeline editor already open, some videos caused an infinite loop while MMP waited for the duration to become available. This is a mistake. The issue was that, occasionally, the duration is returned as a decimal rather than as an integer, which is not clear from the MediaInfo.dll documentation. Consequently, tryStrToInt was failing and returning zero. Changing to getting the duration from MediaInfo instead of waiting for MPV was the correct move; however that came with its own problem. reInitTimeline should be good now.