reduced list of "dirty" characters
In practice, it's only a single quote in a filename that screws up ffmpeg's parsing of the .seg file. The overly aggressive implementation of mmpIsEditFriendly meant that MMP refused to allow editing of file names that contain diacritics, cyrillic or chinese characters, etc. That was never the intention but the implementation was very Latin-alphabet-centric. This has been corrected. The user can still override the list of "dirty" characters in the .conf file if more than a single quote causes them problems. Ctrl-Shift-R to cleanup the file name has also been brought into line.