Minor tweak to Timeline Editing position of main window
If the main window is centered when Timeline Editing is entered with Ctrl-[E], it only gets shifted left if any part of the StreamList / Segments window would be beyond the right edge of the screen. If not, then editing takes place centrally which is more visually appealing to the user.