Copy source file to fix ffmpeg export errors
If ffmpeg has problems with stream IDs during export from the Timeline Editor, simply copying the entire file using ffmpeg can re-index the stream IDs allowing the export to succeed . Ctrl-Shift changes the Export button to Copy, which loads a copy of the video into the editor, together with a copy of the .mmp file. The video is positioned at the correct timestamp.